購入の簡易化機能の実装
