--- app/helpers/purchases_helper.rb
@@ -1,2 +1,0 @@
-module PurchasesHelper
-end
