+++ app/models/purchase.rb
@@ -1,5 +1,4 @@
 class Purchase < ApplicationRecord
   belongs_to :user
   belongs_to :item
-  has_one :shipping_adress
 end
