+++ app/views/items/show.html.erb
@@ -31,11 +31,11 @@
     <% if Purchase.find_by(item_id: @item.id).nil? %>
       <% if user_signed_in? %>
         <% if @item.user == current_user %>
-          <%= link_to "商品の編集", edit_item_path(@item.id), method: :get, class: "item-red-btn" %>
+          <%= link_to "商品の編集", edit_item_path(@item.id), method: :post, class: "item-red-btn" %>
           <p class="or-text">or</p>
           <%= link_to "削除", item_path(@item.id), method: :delete, class:"item-destroy" %>
         <% else%>
-          <%= link_to "購入画面に進む", item_purchases_path(@item.id),class:"item-red-btn"%>
+          <%= link_to "購入する", purchase_item_path(@item.id), method: :post, class:"item-red-btn"%>
         <% end %>
       <% end %>
     <% end %>
