+++ config/routes.rb
@@ -3,7 +3,7 @@ Rails.application.routes.draw do
   devise_for :users
   root to: "items#index"
   resources :items do
-    resources :purchases, only: [:index, :create]
+    post 'purchase', on: :member
   end
   resources :users, only: [:edit, :update]
   resources :cards, only: [:new, :create]
