--- spec/requests/purchases_spec.rb
@@ -1,7 +1,0 @@
-require 'rails_helper'
-
-RSpec.describe 'Purchases', type: :request do
-  describe 'GET /index' do
-    pending "add some examples (or delete) #{__FILE__}"
-  end
-end
